Desktop/Algorizz/requirements.txt
gradio
langchain
langchain-community
langchain-openai
faiss-cpu
sentence-transformers
pypdf
openai
python-dotenv